requirements.txt
requests
# psycopg2
discord.py
pafy
# icecream
python-dotenv
youtube-dl
asgiref
pynacl
opuslib==3.0.1
# git+https://github.com/tamland/python-tidal@0.7.x#egg=tidalapi